refactor(p2): extract Models tab + cards into sibling page (stage 3 of 3)
SettingProviderModelPage + ModelList + ProviderSetting.modelListRequestKey +
ProviderModelListRequestKey (data class) + AddModelButton + ModelPicker +
ModelCard moved from SettingProviderDetailPage.kt (L230-943 of the Stage-2-
trimmed coordinator) into new sibling SettingProviderModelTabPage.kt.

Byte-equal: bodies and signatures unchanged.
Visibility widened: SettingProviderModelPage private -> internal (hub
SettingProviderDetailPage still calls it cross-file).
Other 6 decls (ModelList, modelListRequestKey, ProviderModelListRequestKey,
AddModelButton, ModelPicker, ModelCard) stay private — all callers live
inside this same sibling.

Coordinator strip: 943 -> 169 (-774 with import cleanup).
Imports trimmed: 58 this-refactor orphans across material3/foundation/
runtime/ui/ai/components clusters (all confirmed >=1 body ref in 67e7e442
baseline, 0 in trimmed coordinator). Notably includes setValue + plus
(implicit usage in extracted body via 3 var-by sites + PaddingValues+
operator — all moved to sibling, so coord no longer needs them).
getValue kept — hub still has `val settings by` at L115.

Pre-existing dead imports preserved per CLAUDE.md §3 (10 total):
Connect, height, LinearWavyProgressIndicator, CoroutineScope,
TextGenerationParams, defaultCodexOAuthModelList, UIMessage, ModelSelector,
extendColors, UiState.

Final coordinator state: 169 lines (down from 1705 baseline, -90.1%).
Just SettingProviderDetailPage hub composable + imports + EOF clean.

Baseline 3/467 unit tests preserved.
3-variant assemble: Debug + Notion + Refactortest all BUILD SUCCESSFUL.
