+++ app/src/main/java/me/rerere/rikkahub/ui/pages/setting/SettingProviderModelTabPage.kt
@@ -1,0 +1,797 @@
+package me.rerere.rikkahub.ui.pages.setting
+
+import me.rerere.hugeicons.HugeIcons
+import me.rerere.hugeicons.stroke.Package01
+import me.rerere.hugeicons.stroke.ArrowDown01
+import me.rerere.hugeicons.stroke.Add01
+import me.rerere.hugeicons.stroke.Edit01
+import me.rerere.hugeicons.stroke.Delete01
+import me.rerere.hugeicons.stroke.Cancel01
+import androidx.compose.animation.AnimatedVisibility
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.FlowRow
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.imePadding
+import androidx.compose.foundation.layout.offset
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
+import androidx.compose.foundation.lazy.rememberLazyListState
+import androidx.compose.material3.Badge
+import androidx.compose.material3.BadgedBox
+import androidx.compose.material3.Button
+import androidx.compose.material3.Card
+import androidx.compose.material3.FilledIconButton
+import androidx.compose.material3.FloatingToolbarDefaults.ScreenOffset
+import androidx.compose.material3.FloatingToolbarDefaults.floatingToolbarVerticalNestedScroll
+import androidx.compose.material3.HorizontalFloatingToolbar
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.ModalBottomSheet
+import androidx.compose.material3.OutlinedCard
+import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Surface
+import androidx.compose.material3.SwipeToDismissBox
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
+import androidx.compose.material3.rememberModalBottomSheetState
+import androidx.compose.material3.rememberSwipeToDismissBoxState
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.produceState
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.graphicsLayer
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.util.fastFilter
+import kotlinx.coroutines.launch
+import me.rerere.ai.provider.Model
+import me.rerere.ai.provider.OpenAIAuthMode
+import me.rerere.ai.provider.ProviderManager
+import me.rerere.ai.provider.ProviderSetting
+import me.rerere.ai.provider.providers.isCodexOAuthReviewModel
+import me.rerere.ai.registry.ModelRegistry
+import me.rerere.rikkahub.R
+import me.rerere.rikkahub.ui.components.ai.ModelAbilityTag
+import me.rerere.rikkahub.ui.components.ai.ModelModalityTag
+import me.rerere.rikkahub.ui.components.ai.ModelTypeTag
+import me.rerere.rikkahub.ui.components.ui.AutoAIIcon
+import me.rerere.rikkahub.ui.components.ui.Tag
+import me.rerere.rikkahub.ui.components.ui.TagType
+import me.rerere.rikkahub.ui.hooks.useEditState
+import me.rerere.rikkahub.utils.plus
+import org.koin.compose.koinInject
+import sh.calvin.reorderable.ReorderableItem
+import sh.calvin.reorderable.rememberReorderableLazyListState
+
+@Composable
+internal fun SettingProviderModelPage(
+    provider: ProviderSetting,
+    onEdit: (ProviderSetting) -> Unit
+) {
+    ModelList(
+        providerSetting = provider,
+        onUpdateProvider = onEdit
+    )
+}
+
+@Composable
+private fun ModelList(
+    providerSetting: ProviderSetting,
+    onUpdateProvider: (ProviderSetting) -> Unit
+) {
+    val providerManager = koinInject<ProviderManager>()
+    val requestKey = remember(providerSetting) { providerSetting.modelListRequestKey() }
+    val modelList by produceState(emptyList(), requestKey) {
+        runCatching {
+            println("loading models...")
+            value = providerManager.getProviderByType(providerSetting)
+                .listModels(providerSetting)
+                .sortedBy { it.modelId }
+                .toList()
+        }.onFailure {
+            it.printStackTrace()
+        }
+    }
+    LaunchedEffect(providerSetting, modelList) {
+        if (
+            providerSetting is ProviderSetting.OpenAI &&
+            providerSetting.authMode == OpenAIAuthMode.CODEX_OAUTH &&
+            providerSetting.models.any { it.isCodexOAuthReviewModel() }
+        ) {
+            // One-shot cleanup: drop review variants if any sneaked into the user's selection
+            // from an old Codex response. We do NOT touch the rest — and we explicitly DO NOT
+            // auto-fill on `models.isEmpty()` (an earlier version did, which made the
+            // "deselect all" gesture useless because the next compose pass refilled it).
+            // Keeping the user's empty selection empty is correct: the candidates live in
+            // `modelList` / the available-models sheet, and the user picks from there.
+            val filtered = providerSetting.models.filterNot { it.isCodexOAuthReviewModel() }
+            onUpdateProvider(providerSetting.copy(models = filtered))
+        }
+    }
+    var expanded by rememberSaveable { mutableStateOf(true) }
+    val lazyListState = rememberLazyListState()
+    val reorderableLazyListState = rememberReorderableLazyListState(lazyListState) { from, to ->
+        onUpdateProvider(providerSetting.moveMove(from.index, to.index))
+    }
+
+    Box(modifier = Modifier.fillMaxSize()) {
+        LazyColumn(
+            modifier = Modifier
+                .fillMaxSize()
+                .floatingToolbarVerticalNestedScroll(
+                    expanded = expanded,
+                    onExpand = { expanded = true },
+                    onCollapse = { expanded = false },
+                ),
+            contentPadding = PaddingValues(16.dp) + PaddingValues(bottom = 128.dp),
+            horizontalAlignment = Alignment.CenterHorizontally,
+            verticalArrangement = Arrangement.spacedBy(8.dp),
+            state = lazyListState
+        ) {
+            // 模型列表
+            if (providerSetting.models.isEmpty()) {
+                item {
+                    Column(
+                        modifier = Modifier
+                            .fillParentMaxHeight(0.8f)
+                            .fillMaxWidth(),
+                        horizontalAlignment = Alignment.CenterHorizontally,
+                        verticalArrangement = Arrangement.Center
+                    ) {
+                        Text(
+                            text = stringResource(R.string.setting_provider_page_no_models),
+                            style = MaterialTheme.typography.bodyLarge,
+                            color = MaterialTheme.colorScheme.onSurfaceVariant
+                        )
+                        Text(
+                            text = stringResource(R.string.setting_provider_page_add_models_hint),
+                            style = MaterialTheme.typography.bodyMedium,
+                            color = MaterialTheme.colorScheme.onSurfaceVariant.copy(alpha = 0.7f)
+                        )
+                    }
+                }
+            } else {
+                items(providerSetting.models, key = { it.id }) { item ->
+                    ReorderableItem(
+                        state = reorderableLazyListState,
+                        key = item.id
+                    ) { isDragging ->
+                        ModelCard(
+                            model = item,
+                            onDelete = {
+                                onUpdateProvider(providerSetting.delModel(item))
+                            },
+                            onEdit = { editedModel ->
+                                onUpdateProvider(providerSetting.editModel(editedModel))
+                            },
+                            parentProvider = providerSetting,
+                            modifier = Modifier
+                                .longPressDraggableHandle()
+                                .graphicsLayer {
+                                    if (isDragging) {
+                                        scaleX = 1.05f
+                                        scaleY = 1.05f
+                                    } else {
+                                        scaleX = 1f
+                                        scaleY = 1f
+                                    }
+                                },
+                        )
+                    }
+                }
+            }
+        }
+        HorizontalFloatingToolbar(
+            expanded = expanded,
+            modifier = Modifier
+                .align(Alignment.BottomCenter)
+                .offset(y = -ScreenOffset),
+        ) {
+            AddModelButton(
+                models = modelList,
+                selectedModels = providerSetting.models,
+                onAddModel = {
+                    onUpdateProvider(providerSetting.addModel(it))
+                },
+                onRemoveModel = {
+                    onUpdateProvider(providerSetting.delModel(it))
+                },
+                expanded = expanded,
+                parentProvider = providerSetting,
+                onUpdateProvider = onUpdateProvider
+            )
+        }
+    }
+}
+
+private fun ProviderSetting.modelListRequestKey(): ProviderModelListRequestKey {
+    return when (this) {
+        is ProviderSetting.OpenAI -> ProviderModelListRequestKey(
+            type = "openai",
+            id = id.toString(),
+            credentialsHash = apiKey.hashCode(),
+            baseUrl = baseUrl,
+            authMode = authMode.name,
+            extra = "$chatCompletionsPath|$useResponseApi|${brand.name}",
+        )
+
+        is ProviderSetting.Google -> ProviderModelListRequestKey(
+            type = "google",
+            id = id.toString(),
+            credentialsHash = "$apiKey|$privateKey".hashCode(),
+            baseUrl = baseUrl,
+            authMode = authMode.name,
+            extra = "$vertexAI|$useServiceAccount|$serviceAccountEmail|$location|$projectId",
+        )
+
+        is ProviderSetting.Claude -> ProviderModelListRequestKey(
+            type = "claude",
+            id = id.toString(),
+            credentialsHash = apiKey.hashCode(),
+            baseUrl = baseUrl,
+            authMode = "",
+            extra = promptCaching.toString(),
+        )
+    }
+}
+
+private data class ProviderModelListRequestKey(
+    val type: String,
+    val id: String,
+    val credentialsHash: Int,
+    val baseUrl: String,
+    val authMode: String,
+    val extra: String,
+)
+
+@Composable
+private fun AddModelButton(
+    models: List<Model>,
+    selectedModels: List<Model>,
+    expanded: Boolean,
+    onAddModel: (Model) -> Unit,
+    onRemoveModel: (Model) -> Unit,
+    parentProvider: ProviderSetting,
+    onUpdateProvider: (ProviderSetting) -> Unit
+) {
+    val dialogState = useEditState<Model> { onAddModel(it) }
+    val scope = rememberCoroutineScope()
+
+    Row(
+        horizontalArrangement = Arrangement.spacedBy(16.dp),
+        verticalAlignment = Alignment.CenterVertically,
+    ) {
+        ModelPicker(
+            models = models,
+            selectedModels = selectedModels,
+                onModelSelected = { model ->
+                    val inputModalities = ModelRegistry.MODEL_INPUT_MODALITIES.getData(model.modelId)
+                    val outputModalities = ModelRegistry.MODEL_OUTPUT_MODALITIES.getData(model.modelId)
+                    val abilities = ModelRegistry.MODEL_ABILITIES.getData(model.modelId)
+                    val contextWindowTokens = ModelRegistry.MODEL_CONTEXT_WINDOW.getData(model.modelId)
+                    onAddModel(
+                        model.copy(
+                            inputModalities = inputModalities,
+                            outputModalities = outputModalities,
+                            abilities = abilities,
+                            contextWindowTokens = contextWindowTokens,
+                        )
+                    )
+                },
+            onModelDeselected = { model ->
+                onRemoveModel(model)
+            },
+            onAllModelSelected = {
+                onUpdateProvider(
+                    parentProvider.copyProvider(
+                        models = parentProvider.models + it.filterNot { model ->
+                            parentProvider is ProviderSetting.OpenAI &&
+                                parentProvider.authMode == OpenAIAuthMode.CODEX_OAUTH &&
+                                model.isCodexOAuthReviewModel()
+                        }.filter { model ->
+                            parentProvider.models.none { existing -> existing.modelId == model.modelId }
+                        }.map { model ->
+                            model.copy(
+                                inputModalities = ModelRegistry.MODEL_INPUT_MODALITIES.getData(model.modelId),
+                                outputModalities = ModelRegistry.MODEL_OUTPUT_MODALITIES.getData(model.modelId),
+                                abilities = ModelRegistry.MODEL_ABILITIES.getData(model.modelId),
+                                contextWindowTokens = ModelRegistry.MODEL_CONTEXT_WINDOW.getData(model.modelId),
+                            )
+                        }
+                    )
+                )
+            },
+            onAllModelDeselected = { filteredModels ->
+                onUpdateProvider(
+                    parentProvider.copyProvider(
+                        models = parentProvider.models.filter { model ->
+                            filteredModels.none { filtered -> filtered.modelId == model.modelId }
+                        }
+                    )
+                )
+            }
+        )
+
+        Button(
+            onClick = {
+                dialogState.open(Model())
+            }
+        ) {
+            Row(
+                modifier = Modifier,
+                horizontalArrangement = Arrangement.Center,
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                Icon(
+                    HugeIcons.Add01,
+                    contentDescription = stringResource(R.string.setting_provider_page_add_model)
+                )
+                AnimatedVisibility(expanded) {
+                    Spacer(modifier = Modifier.size(8.dp))
+                    Text(
+                        stringResource(R.string.setting_provider_page_add_new_model),
+                        style = MaterialTheme.typography.bodyLarge
+                    )
+                }
+            }
+        }
+    }
+
+    if (dialogState.isEditing) {
+        dialogState.currentState?.let { modelState ->
+            val sheetState = rememberModalBottomSheetState(skipPartiallyExpanded = true)
+            ModalBottomSheet(
+                onDismissRequest = {
+                    dialogState.dismiss()
+                },
+                sheetState = sheetState,
+                sheetGesturesEnabled = false,
+                dragHandle = {
+                    IconButton(
+                        onClick = {
+                            scope.launch {
+                                sheetState.hide()
+                                dialogState.dismiss()
+                            }
+                        }
+                    ) {
+                        Icon(HugeIcons.ArrowDown01, null)
+                    }
+                }
+            ) {
+                Column(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .fillMaxHeight(0.95f)
+                        .padding(16.dp),
+                    verticalArrangement = Arrangement.spacedBy(16.dp),
+                    horizontalAlignment = Alignment.CenterHorizontally,
+                ) {
+                    Text(
+                        text = stringResource(R.string.setting_provider_page_add_model),
+                        style = MaterialTheme.typography.titleLarge
+                    )
+                    Column(
+                        verticalArrangement = Arrangement.spacedBy(8.dp),
+                        modifier = Modifier
+                            .weight(1f)
+                            .fillMaxWidth()
+                    ) {
+                        ModelSettingsForm(
+                            model = modelState,
+                            onModelChange = { dialogState.currentState = it },
+                            isEdit = false,
+                            parentProvider = parentProvider
+                        )
+                    }
+
+                    Row(
+                        modifier = Modifier.fillMaxWidth(),
+                        horizontalArrangement = Arrangement.spacedBy(8.dp, Alignment.End),
+                    ) {
+                        TextButton(
+                            onClick = {
+                                dialogState.dismiss()
+                            },
+                        ) {
+                            Text(stringResource(R.string.cancel))
+                        }
+                        TextButton(
+                            onClick = {
+                                if (modelState.modelId.isNotBlank() && modelState.displayName.isNotBlank()) {
+                                    dialogState.confirm()
+                                }
+                            },
+                        ) {
+                            Text(stringResource(R.string.setting_provider_page_add))
+                        }
+                    }
+                }
+            }
+        }
+    }
+}
+
+@Composable
+private fun ModelPicker(
+    models: List<Model>,
+    selectedModels: List<Model>,
+    onModelSelected: (Model) -> Unit,
+    onModelDeselected: (Model) -> Unit,
+    onAllModelSelected: (List<Model>) -> Unit,
+    onAllModelDeselected: (List<Model>) -> Unit
+) {
+    var showModal by remember { mutableStateOf(false) }
+    if (showModal) {
+        ModalBottomSheet(
+            onDismissRequest = { showModal = false },
+            sheetState = rememberModalBottomSheetState(skipPartiallyExpanded = true),
+        ) {
+            var filterText by remember { mutableStateOf("") }
+            val filterKeywords = filterText.split(" ").filter { it.isNotBlank() }
+            val filteredModels = models.fastFilter {
+                if (filterKeywords.isEmpty()) {
+                    true
+                } else {
+                    filterKeywords.all { keyword ->
+                        it.modelId.contains(keyword, ignoreCase = true) ||
+                            it.displayName.contains(keyword, ignoreCase = true)
+                    }
+                }
+            }
+            Column(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .fillMaxHeight(0.9f)
+                    .padding(8.dp)
+                    .imePadding(),
+                verticalArrangement = Arrangement.spacedBy(4.dp)
+            ) {
+                // 标题栏和添加所有按钮
+                Row(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .padding(horizontal = 8.dp, vertical = 4.dp),
+                    horizontalArrangement = Arrangement.SpaceBetween,
+                    verticalAlignment = Alignment.CenterVertically
+                ) {
+                    Text(
+                        text = stringResource(R.string.setting_provider_page_avaliable_models),
+                        style = MaterialTheme.typography.titleMedium
+                    )
+
+                    val unselectedCount = filteredModels.count { model ->
+                        selectedModels.none { it.modelId == model.modelId }
+                    }
+
+                    TextButton(
+                        onClick = {
+                            if (unselectedCount > 0) {
+                                onAllModelSelected(filteredModels)
+                            } else {
+                                onAllModelDeselected(filteredModels)
+                            }
+                        },
+                    ) {
+                        Text(
+                            if (unselectedCount > 0) stringResource(
+                                R.string.setting_provider_page_select_all,
+                                unselectedCount
+                            ) else stringResource(R.string.setting_provider_page_deselect_models)
+                        )
+                    }
+                }
+
+                LazyColumn(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .weight(1f),
+                    verticalArrangement = Arrangement.spacedBy(8.dp),
+                    contentPadding = PaddingValues(8.dp),
+                ) {
+                    items(filteredModels) {
+                        Card {
+                            Row(
+                                verticalAlignment = Alignment.CenterVertically,
+                                horizontalArrangement = Arrangement.spacedBy(
+                                    8.dp
+                                ),
+                                modifier = Modifier
+                                    .fillMaxWidth()
+                                    .padding(8.dp),
+                            ) {
+                                AutoAIIcon(
+                                    it.modelId,
+                                    Modifier.size(32.dp)
+                                )
+                                Column(
+                                    verticalArrangement = Arrangement.spacedBy(
+                                        4.dp
+                                    ),
+                                    modifier = Modifier.weight(1f),
+                                ) {
+                                    Text(
+                                        text = it.modelId,
+                                        style = MaterialTheme.typography.titleSmall,
+                                    )
+
+                                    Row(
+                                        verticalAlignment = Alignment.CenterVertically,
+                                        horizontalArrangement = Arrangement.spacedBy(2.dp)
+                                    ) {
+                                        val modelMeta = remember(it) {
+                                            it.copy(
+                                                inputModalities = ModelRegistry.MODEL_INPUT_MODALITIES.getData(it.modelId),
+                                                outputModalities = ModelRegistry.MODEL_OUTPUT_MODALITIES.getData(it.modelId),
+                                                abilities = ModelRegistry.MODEL_ABILITIES.getData(it.modelId),
+                                                contextWindowTokens = ModelRegistry.MODEL_CONTEXT_WINDOW.getData(it.modelId),
+                                            )
+                                        }
+                                        ModelModalityTag(
+                                            model = modelMeta,
+                                        )
+                                        ModelAbilityTag(
+                                            model = modelMeta,
+                                        )
+                                    }
+                                }
+                                IconButton(
+                                    onClick = {
+                                        if (selectedModels.any { model -> model.modelId == it.modelId }) {
+                                            // 从selectedModels中计算出要删除的model，因为删除需要id匹配，而不是ModelId
+                                            onModelDeselected(selectedModels.firstOrNull { model -> model.modelId == it.modelId }
+                                                ?: it)
+                                        } else {
+                                            onModelSelected(it)
+                                        }
+                                    }
+                                ) {
+                                    if (selectedModels.any { model -> model.modelId == it.modelId }) {
+                                        Icon(HugeIcons.Cancel01, null)
+                                    } else {
+                                        Icon(HugeIcons.Add01, null)
+                                    }
+                                }
+                            }
+                        }
+                    }
+                }
+                OutlinedTextField(
+                    value = filterText,
+                    onValueChange = {
+                        filterText = it
+                    },
+                    label = { Text(stringResource(R.string.setting_provider_page_filter_placeholder)) },
+                    modifier = Modifier.fillMaxWidth(),
+                    placeholder = {
+                        Text(stringResource(R.string.setting_provider_page_filter_example))
+                    },
+                )
+            }
+        }
+    }
+    BadgedBox(
+        badge = {
+            if (models.isNotEmpty()) {
+                Badge {
+                    Text(models.size.toString())
+                }
+            }
+        }
+    ) {
+        IconButton(
+            onClick = {
+                showModal = true
+            }
+        ) {
+            Icon(HugeIcons.Package01, null)
+        }
+    }
+}
+
+@Composable
+private fun ModelCard(
+    model: Model,
+    modifier: Modifier = Modifier,
+    onDelete: () -> Unit,
+    onEdit: (Model) -> Unit,
+    parentProvider: ProviderSetting
+) {
+    val dialogState = useEditState<Model> {
+        onEdit(it)
+    }
+    val swipeToDismissBoxState = rememberSwipeToDismissBoxState()
+    val scope = rememberCoroutineScope()
+
+    if (dialogState.isEditing) {
+        dialogState.currentState?.let { editingModel ->
+            val sheetState = rememberModalBottomSheetState(skipPartiallyExpanded = true)
+            ModalBottomSheet(
+                onDismissRequest = {
+                    dialogState.dismiss()
+                },
+                sheetState = sheetState,
+                sheetGesturesEnabled = false,
+                dragHandle = null,
+            ) {
+                Column(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .fillMaxHeight(0.95f)
+                        .padding(16.dp),
+                    verticalArrangement = Arrangement.spacedBy(16.dp),
+                    horizontalAlignment = Alignment.CenterHorizontally,
+                ) {
+                    Box(
+                        modifier = Modifier.fillMaxWidth(),
+                    ) {
+                        IconButton(
+                            onClick = {
+                                scope.launch {
+                                    sheetState.hide()
+                                    dialogState.dismiss()
+                                }
+                            },
+                            modifier = Modifier.align(Alignment.CenterStart)
+                        ) {
+                            Icon(HugeIcons.Cancel01, null)
+                        }
+                        Text(
+                            text = stringResource(R.string.setting_provider_page_edit_model),
+                            style = MaterialTheme.typography.titleLarge,
+                            modifier = Modifier.align(Alignment.Center),
+                        )
+                    }
+                    Column(
+                        verticalArrangement = Arrangement.spacedBy(8.dp),
+                        modifier = Modifier
+                            .weight(1f)
+                            .fillMaxWidth()
+                    ) {
+                        ModelSettingsForm(
+                            model = editingModel,
+                            onModelChange = { dialogState.currentState = it },
+                            isEdit = true,
+                            parentProvider = parentProvider
+                        )
+                    }
+
+                    Row(
+                        modifier = Modifier.fillMaxWidth(),
+                        horizontalArrangement = Arrangement.spacedBy(8.dp, Alignment.End),
+                    ) {
+                        TextButton(
+                            onClick = {
+                                dialogState.dismiss()
+                            },
+                        ) {
+                            Text(stringResource(R.string.cancel))
+                        }
+                        TextButton(
+                            onClick = {
+                                if (editingModel.displayName.isNotBlank()) {
+                                    dialogState.confirm()
+                                }
+                            },
+                        ) {
+                            Text(stringResource(R.string.confirm))
+                        }
+                    }
+                }
+            }
+        }
+    }
+
+    SwipeToDismissBox(
+        state = swipeToDismissBoxState,
+        backgroundContent = {
+            Row(
+                modifier = Modifier
+                    .fillMaxSize()
+                    .padding(8.dp),
+                horizontalArrangement = Arrangement.spacedBy(8.dp, Alignment.End),
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                IconButton(
+                    onClick = {
+                        scope.launch {
+                            swipeToDismissBoxState.reset()
+                        }
+                    }
+                ) {
+                    Icon(HugeIcons.Cancel01, null)
+                }
+                FilledIconButton(
+                    onClick = {
+                        scope.launch {
+                            onDelete()
+                            swipeToDismissBoxState.reset()
+                        }
+                    }
+                ) {
+                    Icon(
+                        HugeIcons.Delete01,
+                        contentDescription = stringResource(R.string.chat_page_delete)
+                    )
+                }
+            }
+        },
+        enableDismissFromStartToEnd = false,
+        gesturesEnabled = true,
+        modifier = modifier
+    ) {
+        OutlinedCard {
+            Row(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(horizontal = 16.dp, vertical = 8.dp),
+                horizontalArrangement = Arrangement.spacedBy(16.dp),
+                verticalAlignment = Alignment.CenterVertically,
+            ) {
+                Surface(
+                    color = MaterialTheme.colorScheme.secondaryContainer,
+                    shape = MaterialTheme.shapes.small,
+                ) {
+                    AutoAIIcon(
+                        name = model.modelId,
+                        modifier = Modifier.size(36.dp),
+                    )
+                }
+                Column(
+                    modifier = Modifier.weight(1f),
+                    verticalArrangement = Arrangement.spacedBy(4.dp),
+                ) {
+                    Text(
+                        text = model.displayName,
+                        style = MaterialTheme.typography.titleSmall,
+                        maxLines = 2,
+                        overflow = TextOverflow.Ellipsis,
+                    )
+                    FlowRow(
+                        horizontalArrangement = Arrangement.spacedBy(4.dp),
+                        verticalArrangement = Arrangement.spacedBy(4.dp),
+                    ) {
+                        if (model.providerOverwrite != null) {
+                            Tag(type = TagType.INFO) {
+                                Text(
+                                    model.providerOverwrite?.javaClass?.simpleName ?: model.providerOverwrite?.name
+                                    ?: "ProviderOverwrite"
+                                )
+                            }
+                        }
+                        ModelTypeTag(model = model)
+                        ModelModalityTag(model = model)
+                        ModelAbilityTag(model = model)
+                    }
+                }
+
+                // Edit button
+                IconButton(
+                    onClick = {
+                        dialogState.open(model.copy())
+                    }
+                ) {
+                    Icon(HugeIcons.Edit01, "Edit", modifier = Modifier.size(20.dp))
+                }
+            }
+        }
+    }
+}
